# perf: localize import for `ape console`