### What I did

accidentally did a top level import in https://github.com/ApeWorX/ape/pull/2370
which caused a significant enough slow down to all of `ape --help`


```
ape --help  1.30s user 0.17s system 60% cpu 2.425 total
```


```
ape --help  0.95s user 0.13s system 57% cpu 1.887 total
```

### How I did it

<!-- Discuss the thought process behind the change -->

### How to verify it

<!-- Discuss any methods that should be used to verify the change -->

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated